arm/arm64: irq: drop "IRQx no longer affine to CPUn" print to pr_debug()
On devices where hotplug is used for power savings or thermal
mitigation, it's not uncommon for hotplug events to be relatively
frequent and their latency to matter.

If a serial console is enabled, the "IRQ no longer affine to CPU"
informational prints can be one of the most significant contributors
to hotplug latency. For example, on msm8994, 10 lines of this format
are normally printed on each hotplug. At 115200/8/N/1 baud, that
printing takes a minimum of 42.5ms.

On production devices where a serial console is not enabled, the time
takes is negligible (a few microseconds at most), but the prints do
still clutter the logs and are worth silencing since they are
typically of no concern.

Change-Id: I4ee718c48893b03eda98b73ffcc791641435dc62
Signed-off-by: Matt Wagantall <mwagantall@cyngn.com>
(cherry picked from commit 7ff1a1c716d2d01c4653377e225403b204f0cb72)
Signed-off-by: Albert I <kras@raphielgang.org>
Signed-off-by: dndxtz <dindascape@gmail.com>